requirements.txt
# PyTorch with CUDA support - install separately first:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
pyyaml>=6.0
tqdm>=4.65.0
fastdtw>=0.3.4